Closes #21 

다음 사항 구현했습니다.
- 전공 필수, 교양 필수(채플 포함), 전공 선택(사용자가 선택한 것만) 이름으로 조회하기
- 3가지 분야의 과목들을 가지고, 시간표가 성립 가능한 모든 경우의 수 생성(성립이 되는 경우가 하나도 없다면 오류 반환)

+++ src/main/kotlin/com/yourssu/soongpt/domain/rating/implement/Rating.kt
@@ -4,7 +4,7 @@ class Rating(
     val id: Long? = null,
     val courseName: String,
     val professorName: String,
-    val star: Int,
+    val star: Double,
     var point: Double = 50.0,
 ) {
 }
